+++ src/core/AiTone.tsx
@@ -1,18 +1,15 @@
-/* eslint-disable react-hooks/rules-of-hooks */
 import { __ } from '@wordpress/i18n';
 import { verse } from '@wordpress/icons';
 import { addFilter } from '@wordpress/hooks';
 import { BlockControls } from '@wordpress/block-editor';
 import { select, dispatch } from '@wordpress/data';
-import { Fragment, useState, useEffect } from '@wordpress/element';
 import { ToolbarGroup, ToolbarDropdownMenu } from '@wordpress/components';
 import apiFetch from '@wordpress/api-fetch';
 
-import { selectProps, selectBlockProps } from '../utils/types';
+import { selectProps, selectBlockProps, noticeProps } from '../utils/types';
 import { getAllowedBlocks, getBlockMenuOptions } from '../utils';
 import { editorStore, noticesStore, blockEditorStore } from '../utils/store';
 
-import Toast from '../components/Toast';
 import '../styles/app.scss';
 
 /**
@@ -31,9 +28,6 @@ export const filterBlockTypesWithAI = ( settings: any ): object => {
 	}
 
 	settings.edit = ( props: any ) => {
-		const [ tone, setTone ] = useState( '' );
-		const [ isLoading, setIsLoading ] = useState( false );
-
 		/**
 		 * Get Tone Params.
 		 *
@@ -66,12 +60,26 @@ export const filterBlockTypesWithAI = ( settings: any ): object => {
 		 */
 		const getTone = async ( newTone: string ): Promise< void > => {
 			const { postId, blockId, blockContent } = getToneParams();
-			const { createErrorNotice } = dispatch( noticesStore ) as any;
+			const { createNotice, createErrorNotice, removeNotice } = dispatch(
+				noticesStore
+			) as any;
 			const { updateBlockAttributes } = dispatch(
 				blockEditorStore
 			) as any;
+			const { getNotices } = select( noticesStore ) as noticeProps;
 
-			setIsLoading( true );
+			createNotice(
+				'success',
+				__(
+					'AI is generating text, please hold on for a bit.',
+					'ai-plus-block-editor'
+				),
+				{
+					isDismissible: true,
+					id: 'apbe-tone-success',
+					type: 'snackbar',
+				}
+			);
 
 			try {
 				const aiTone: string = await apiFetch( {
@@ -95,7 +103,9 @@ export const filterBlockTypesWithAI = ( settings: any ): object => {
 					limit++;
 				}, 5 );
 
-				setIsLoading( false );
+				getNotices().forEach( ( notice: any ) =>
+					removeNotice( notice.id )
+				);
 			} catch ( e ) {
 				createErrorNotice( e.message, {
 					type: 'snackbar',
@@ -104,34 +114,19 @@ export const filterBlockTypesWithAI = ( settings: any ): object => {
 			}
 		};
 
-		useEffect( () => {
-			if ( tone ) {
-				getTone( tone );
-			}
-			// eslint-disable-next-line react-hooks/exhaustive-deps
-		}, [ tone ] );
-
 		return (
-			<Fragment>
-				<Toast
-					isInEditor={ true }
-					message={ __(
-						'AI is generating text, please hold on for a bit.',
-						'ai-plus-block-editor'
-					) }
-					isLoading={ isLoading }
-				/>
+			<>
 				<BlockControls>
 					<ToolbarGroup>
 						<ToolbarDropdownMenu
 							icon={ verse }
 							label={ __( 'AI + Block Editor' ) }
-							controls={ getBlockMenuOptions( setTone ) }
+							controls={ getBlockMenuOptions( getTone ) }
 						/>
 					</ToolbarGroup>
 				</BlockControls>
 				{ edit( props ) }
-			</Fragment>
+			</>
 		);
 	};