+++ src/utils/index.tsx
@@ -9,46 +9,47 @@ type DropdownOption = {
 /**
  * Get Options.
  *
- * This function filters the Block Control Options
+ * This function filters the Block Menu AI Options
  * available when the user selects a block.
  *
  * @since 1.1.0
  *
- * @param {Function} setTone Subscribe to setTone setter method.
+ * @param {Function} getTone Function to get the tone.
  * @return {DropdownOption[]} Dropdown options.
  */
-export const getBlockMenuOptions = ( setTone: Function ): DropdownOption[] => {
+export const getBlockMenuOptions = ( getTone: Function ): DropdownOption[] => {
 	const menu = [];
 
-	const options = {
+	/**
+	 * Filter Menu Options.
+	 *
+	 * By default the passed object should contain
+	 * menu options.
+	 *
+	 * @since 1.1.0
+	 * @since 1.5.0 Pass in menu options before returning menu.
+	 *
+	 * @param {Object} menu Menu options.
+	 * @return {Object}
+	 */
+	const options = applyFilters( 'apbe.blockMenuOptions', {
 		casual: __( 'Use Casual Tone', 'ai-plus-block-editor' ),
 		official: __( 'Use Official Tone', 'ai-plus-block-editor' ),
 		descriptive: __( 'Use Descriptive Tone', 'ai-plus-block-editor' ),
 		narrative: __( 'Use Narrative Tone', 'ai-plus-block-editor' ),
 		aggressive: __( 'Use Aggressive Tone', 'ai-plus-block-editor' ),
-	};
+	} );
 
 	Object.keys( options ).forEach( ( key ) => {
 		menu.push( {
 			title: options[ key ],
 			onClick: () => {
-				setTone( key );
+				getTone( key );
 			},
 		} );
 	} );
 
-	/**
-	 * Filter Menu.
-	 *
-	 * By default the passed option should contain
-	 * menu objects.
-	 *
-	 * @since 1.1.0
-	 *
-	 * @param {DropdownOption[]} menu Menu array containing menu objects.
-	 * @return {DropdownOption[]}
-	 */
-	return applyFilters( 'apbe.blockMenuOptions', menu ) as DropdownOption[];
+	return menu;
 };
 
 /**