+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Feat: Add Social Media Hashtags Feature.
 * Fix: Breaking issue with Tone functionality.
 * Test: Fix Unit tests & update snapshots.
+* Chore: Relocate `apbe.blockMenuOptions` hook.
 
 # 1.4.0
 * Refactor: Adopt useSelect & useDispatch hooks in sidebar components.